BigFloat add inverse method (#51)

+++ include/nnoops/long_arithmetic/unsigned_big_integer.hpp
@@ -341,9 +341,9 @@ struct UBigInteger {
       uint64_t q = el1 / divisor.data[n];
 
       // Test
-      while (q == (BASE + 1) ||
-             (n >= 1 && q * divisor.data[n - 1] >
-                            ((BASE + 1) * r + dividend.data[j + n - 1]))) {
+      while (q == ((uint64_t)BASE + 1) ||
+             (n >= 1 && q * divisor.data[n - 1] > (((uint64_t)BASE + 1) * r +
+                                                   dividend.data[j + n - 1]))) {
         --q;
         r += divisor.data[n];
         if (r > BASE) {
